+++ lib/core/services/data_service/service.dart
@@ -37,7 +37,7 @@ Future<String> voteContentFromPlatform(
   String username,
   String author,
   String permlink,
-  double weight,
+  int weight,
   String? postingKey,
   String? authKey,
   String? token,
