+++ lib/core/services/data_service/web_service.dart
@@ -73,7 +73,7 @@ Future<String> voteContentFromPlatform(
   String username,
   String author,
   String permlink,
-  double weight,
+  int weight,
   String? postingKey,
   String? authKey,
   String? token,
